+++ settings.h
@@ -7,17 +7,17 @@ class settings {
   string line;
 
 public:
-    settings(string path);
-    ~settings();
-    void set_logpath(string path);
-    void set_shortcut(string shortc, string file);
-    string check_shortcut(string str, int& i);
-    void list();
+    settings (string path);
+    ~settings ();
+    void set_logpath (string path);
+    void set_shortcut (string shortc, string file);
+    string check_shortcut (string str, short& i);
+    void list ();
 };
 
-settings::settings ( string path ) {
+settings::settings (string path) {
   string spath = "/root/.dumplog/";
-  int i = system("mkdir -p /root/.dumplog");
+  short i = system("mkdir -p /root/.dumplog");
   if (i) {
     cout << "Unable to read/create settings.\nCheck your rights!" << endl;
     exit(1);
@@ -31,11 +31,11 @@ settings::settings ( string path ) {
   }
 }
 
-settings::~settings() {
+settings::~settings () {
   myfile.close();
 }
 
-void settings::set_logpath ( string path ) {
+void settings::set_logpath (string path) {
   if (path.find_last_of("/") == (path.length()-1)) {      // check if path ends with /
     myfile << path << endl;                                          
   } else {
@@ -45,7 +45,7 @@ void settings::set_logpath ( string path ) {
   cout << "Path has been created." << endl;
 }
 
-void settings::set_shortcut ( string shortc, string file ) {
+void settings::set_shortcut (string shortc, string file) {
   if (check_shortcut(shortc, hlpr) == shortc) {
     myfile.clear();
     myfile.seekp(0, ios::end);
@@ -58,7 +58,7 @@ void settings::set_shortcut ( string shortc, string file ) {
     if (ans[0] == 'y') {
     ifstream infile("/root/.dumplog/shortcuts");
     ofstream outfile("/root/.dumplog/shortcuts_", ios::out);
-    int x = 1;
+    short x = 1;
       while (getline(infile, line)) {
         if (x != hlpr) {
 	  outfile << line << endl;
@@ -73,11 +73,11 @@ void settings::set_shortcut ( string shortc, string file ) {
   }
 }
 
-string settings::check_shortcut ( string str, int& i ) {
+string settings::check_shortcut (string str, short& i) {
   string file = str;
   i = 1;
   while (getline(myfile, line)) {
-    int spl = line.find(",");
+    short spl = line.find(",");
     if (str == line.substr(0, spl) && !line.empty()) {
       file = line.substr(spl+1, (line.length()-spl));
       break;
@@ -87,7 +87,7 @@ string settings::check_shortcut ( string str, int& i ) {
   return file;
 }
 
-void settings::list() {
+void settings::list () {
   while (getline(myfile, line)) {
     cout << line << "\n";
   }
